print whithout loop

+++ 0x13-more_singly_linked_lists/lists.h
@@ -29,6 +29,6 @@ listint_t *insert_nodeint_at_index(listint_t **head, unsigned int idx, int n);
 int delete_nodeint_at_index(listint_t **head, unsigned int index);
 listint_t *reverse_listint(listint_t **head);
 size_t print_listint_safe(const listint_t *head);
-size_t detect_loop(const listint_t *head, const listint_t *current, size_t index);
+int detect_loop(const listint_t *head, const listint_t *current, int index);
 
 #endif
